"Corrected and animated submarine movement in main.js"
This commit fixes a bug where the sub's X position was not updating and the scene not re-rendering due to missing code in the event listener. A conversion from the input event value into a number was also added for correct positioning calculation. Now the submarine properly moves along the x-axis on input and the animation is triggered on each change.