Refactor main.js to improve code readability
The code's structure was rearranged to clarify what each section does. Some parts have been repositioned and others extracted into seperate functions to make it easier to understand. The 'Location' class was moved to the top, and event listeners were moved to the bottom. Moreover, the inline code for the 'sub' y location setting was extracted into a separate function 'setSubLocationY', resulting in more straightforward, easier-to-read code.