Replace depth with height in main dimensions
The logic of setting depth in the main dimensions of the cube was replaced with setting the height. This change was implemented to make the dimension setting more intuitive and to eliminate confusion between depth and height dimensions. Also, a new function was added to set the height from the bottom of the cube if needed.