Add depth control to main cube
This commit includes a function to control the depth of the main cube based on input. Previously, only the height and width of the cube could be controlled by input values, but the depth remained static. Now through a 'setMainDimensionDepth' function, the depth of the main cube can be set and updated as necessary. An event listener also