Update visuals and camera fit to user interactions
This commit improves the user interactive experience in the 3D model scene in the main.js file. I modified the cube edge color to follow the original cube colors, but shifted it closer to white to make it brighter and more noticeable. I also added an automatic camera fit operation right after every interaction that modifies the geometry of the scene - ensuring that the audience, main and sub mirrors positions are correctly updated, and fit into view after each operation. This was achieved by calling the `fitCameraToSelection()`