Add a `Cube` class to replace duplicated logic
Created a `Cube` class and refactored existing code to utilize the new class. This change allows for a more consistent, OOP approach and eliminates redundancies in the old code. Functions for manipulating cube properties, like `changeDimensions` or `setWidth`, are now encapsulated within the `Cube` class. Hence, an overall cleaner codebase is achieved which is easier to maintain and understand.