Refactor main.js and extract constants, page elements
The commit includes extracting constants and page elements within main.js into their own, separate files, for the sake of maintaining consistency in the code structure, as well as enhancing readability. The constants.js file was created to hold constants previously in the main.js file, while the htmlPageElements.js file now holds the HTML elements. This helps improve the overall manageability and neatness of the code.