Refactor subMirror creation and positioning
This commit refactors the way the subMirror object is created and positioned in the scene. Submirror creation is now conditional on the state of the subConfigCheckbox and its positioning has been adjusted. The setPosition method in Cube class has also been enabled to reflect changes. This change simplifies the mirror creation and positioning logic and makes it easier to understand and manage. It also improves code readability by reducing redundancy and moving the responsibility of positioning the subMirror to the Cube class itself.