Remove automatic camera zoom, add a manual reset button
Removed the 'fitCameraToSelection()' function invocations from the input event listeners in main.js. This was causing the view to be zoomed out every time an input changes, which could get tiring for the user who wants to look in detail and keep track of small changes. As a solution, introduced a 'Reset Zoom' button in index.html. Clicking on this button calls the 'fitCameraToSelection()' function and reset the camera view, and this event listener is correspondingly added in main.js.