Refactor orientation logic to use constant mapping
The orientation logic in the camera position setup is modified to use the constant mapping "Orientation". Previously, the orientation was directly referred with string-based cases, which can be prone to errors due to typos and casing issues. Now it's more reliable and maintainable with constant variables. An "Orientation" object is added to the constants file to enable this. A call to `fitCameraToSelection` with "SIDE" orientation is also added in main.js for initial setup.