Update calculations for subs position and audience depth
The way we calculate the position of subs and the depth of the audience has been changed. Instead of using 'arrayBottomHeightInput', we are now using 'subDepthInput' for subs position and 'subDistanceFromCenterInput' for audience depth. This change reflects a better understanding of the physics involved and results in more accurate calculations.