Add GSAP for smoother camera animation
The GSAP (GreenSock Animation Platform) was added as a new dependency to achieve smoother camera movements by interpolating between old and new camera positions. This provides a more natural look and feel to the animation and enhances the user experience. This change affects the camera position setting mechanism in our main.js along with the necessary dependency updates in the package and lock files.