Refactor camera fit function and move gsap import
The `fitCameraToSelection` function was refactored, now they always expect an array as input which simplifies the code by removing the type check and handling for different types. This change will lead to more consistent function usage and less chance of errors due to unexpected inputs.

GSAP import was also moved from `constants.js` to `main.js` to improve code organization because `timeline` from gsap is not a constant but rather used as a local variable within specific function.