Feature update for Ian. Add user interaction check to prevent camera autoscaling
A new variable 'userInteracted' was added to keep track of user interactions with the camera controls. The function 'fitCameraToSelection' now only adapts the camera based on the objects and offset if the user has not yet interacted with the camera controls. This change prevents the camera from autoscaling after a user has already manually controlled the camera. The 'userInteracted' flag is reset to 'false' for the 'resetZoom' event and is set to 'true' when the controls are interacted with.