Update timeline kill rule and input event trigger
Added a condition to kill the timeline if it is active. Correspondingly, a new event listener was added to trigger an 'input' event on window load for all inputs that have a value. This helps in ensuring that any pre-filled inputs are properly registered on page load. Also, changed the id of the container from a generic "container" to a more descriptive "sound-system-model-container". This provides better context to the element and avoids id conflicts. Furthermore, HTML layout has been improved by moving the reset button inside the sound system model container. This makes the layout more consistent and independent, especially with regard to the relative positioning of the elements.